Refactor homebrew family of modules
This reduces duplication by:
- Extracting common code to `module_utils`
- Making homebrew, homebrew_cask depend on base class `HomebrewBase`
